// liwords-ui/src/gen/api/proto/macondo/macondo_pb.d.ts
export declare enum GameEvent_Type {
  /**
   * @generated from enum value: TILE_PLACEMENT_MOVE = 0;
   */
  TILE_PLACEMENT_MOVE = 0,

  /**
   * @generated from enum value: PHONY_TILES_RETURNED = 1;
   */
  PHONY_TILES_RETURNED = 1,

  /**
   * @generated from enum value: PASS = 2;
   */
  PASS = 2,

  /**
   * @generated from enum value: CHALLENGE_BONUS = 3;
   */
  CHALLENGE_BONUS = 3,

  /**
   * @generated from enum value: EXCHANGE = 4;
   */
  EXCHANGE = 4,

  /**
   * @generated from enum value: END_RACK_PTS = 5;
   */
  END_RACK_PTS = 5,

  /**
   * @generated from enum value: TIME_PENALTY = 6;
   */
  TIME_PENALTY = 6,

  /**
   * Only for international rules (or after 6 zeroes)
   *
   * @generated from enum value: END_RACK_PENALTY = 7;
   */
  END_RACK_PENALTY = 7,

  /**
   * Lose a turn for challenging a word incorrectly (only for double
   * challenge)
   *
   * @generated from enum value: UNSUCCESSFUL_CHALLENGE_TURN_LOSS = 8;
   */
  UNSUCCESSFUL_CHALLENGE_TURN_LOSS = 8,

  /**
   * Issue a challenge
   *
   * @generated from enum value: CHALLENGE = 9;
   */
  CHALLENGE = 9,
}

/**
 * @generated from enum macondo.GameEvent.Direction
 */
export declare enum GameEvent_Direction {
  /**
   * @generated from enum value: HORIZONTAL = 0;
   */
  HORIZONTAL = 0,

  /**
   * @generated from enum value: VERTICAL = 1;
   */
  VERTICAL = 1,
}